# api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger CartApi
  version: 1.0.0

servers:
  - url: http://localhost:8191/
paths:
  /cart:
    get:
      summary: Returns  converter a cart.
      operationId: getCurrentCart
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCartItemsView"

    post:
      summary: Create an cart.
      operationId: createCart
      requestBody:
        description: Created cart body
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiProductView"
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
        '400':
          description: The specified product ID is invalid (not a number).
        '404':
          description: A product with the specified ID was not found.
        default:
          description: Unexpected error

  /cart,{id}&{productId}:
    get:
      summary: Returns OK.
      operationId: addProductToCart
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: productId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '201':
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 4
                  name:
                    type: string
                    example: Jessica Smith
        '400':
          description: The specified product ID is invalid (not a number).
        '404':
          description: A product with the specified ID was not found.
        default:
          description: Unexpected error
  /cart/{id}:
    put:
      summary: Update cart
      operationId: updateCart
      parameters:
        - name: id
          in: path
          required: true
          description: Cart id
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update cart body
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiProductView"
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCartItem"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    ApiCartItem:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productTitle:
          type: string
        quantity:
          type: integer
          format: int32
        pricePerProduct:
          type: integer
          format: int32
        price:
          type: integer
          format: int32

    ApiProductView:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productTitle:
          type: string
        quantity:
          type: integer
          format: int32
        pricePerProduct:
          type: integer
          format: int32
        price:
          type: integer
          format: int32

    ApiCartItemsView:
      type: array
      items:
        $ref: '#/components/schemas/ApiCartItem'
    ApiProductList:
      type: array
      items:
        $ref: '#/components/schemas/ApiProductView'

    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string